// src/app/interfaces/movie.ts
export interface Movie {
  poster_path: string|null;
  adult: boolean;
  overview: string;
  release_date: string;
  genre_ids: number[];
  id: string|number; // integer
  original_title: string;
  original_language: string;
  title: string;
  backdrop_path: string|null;
  popularity: number;
  vote_count: string|number; // integer
  video: boolean;
  vote_average: number;
}